+++ client/src/main/java/com/influxdb/client/internal/InfluxDBClientImpl.java
@@ -34,7 +34,7 @@ import com.influxdb.client.DashboardsApi;
 import com.influxdb.client.DeleteApi;
 import com.influxdb.client.InfluxDBClient;
 import com.influxdb.client.InfluxDBClientOptions;
-import com.influxdb.client.InvocableScriptsApi;
+import com.influxdb.client.InvokableScriptsApi;
 import com.influxdb.client.LabelsApi;
 import com.influxdb.client.NotificationEndpointsApi;
 import com.influxdb.client.NotificationRulesApi;
@@ -59,7 +59,7 @@ import com.influxdb.client.service.BucketsService;
 import com.influxdb.client.service.ChecksService;
 import com.influxdb.client.service.DashboardsService;
 import com.influxdb.client.service.DeleteService;
-import com.influxdb.client.service.InvocableScriptsService;
+import com.influxdb.client.service.InvokableScriptsService;
 import com.influxdb.client.service.LabelsService;
 import com.influxdb.client.service.NotificationEndpointsService;
 import com.influxdb.client.service.NotificationRulesService;
@@ -244,8 +244,8 @@ public final class InfluxDBClientImpl extends AbstractInfluxDBClient implements 
 
     @NotNull
     @Override
-    public InvocableScriptsApi getInvocableScriptsApi() {
-        return new InvocableScriptsApiImpl(retrofit.create(InvocableScriptsService.class));
+    public InvokableScriptsApi getInvokableScriptsApi() {
+        return new InvokableScriptsApiImpl(retrofit.create(InvokableScriptsService.class));
     }
 
     @Nonnull